# services/identity-intelligence/src/main/java/io/lombardio/identity/portal/api/CustomerPortalInvitationLookupRequest.java
/*
 * Lombardio Source-Available No-Distribution License 1.0
 *
 * Copyright (c) 2026 Benjamin Letzel. All rights reserved.
 *
 * This project is source-available for educational and review purposes only.
 * Redistribution, sublicensing, or commercial use is strictly prohibited.
 *
 * For partnership or cooperation inquiries, please contact the author.
 */
package io.lombardio.identity.portal.api;

import jakarta.validation.constraints.NotBlank;

public record CustomerPortalInvitationLookupRequest(@NotBlank String token) {}